ORM-1344 insert created_at using php time, bypass otp expired if config value is NULL or 0

+++ app/code/local/Bilna/Smsverification/Model/Api2/Verify/Rest/Admin/V1.php
@@ -48,11 +48,12 @@ class Bilna_Smsverification_Model_Api2_Verify_Rest_Admin_V1 extends Bilna_Smsver
             $createdAt = strtotime($OTP['created_at']);
             $currentTime = strtotime(date('Y-m-d H:i:s'));
             $timeOut = Mage::getStoreConfig('bilna/smsverification/timeout');
-            if (($currentTime - $createdAt) > ($timeOut * 60)) {
-                $OTPData->getFirstItem()->delete();
-                $this->_critical('Expired OTP Code');
+            if ((int) $timeout > 0) {
+                if (($currentTime - $createdAt) > ($timeOut * 60)) {
+                    $OTPData->getFirstItem()->delete();
+                    $this->_critical('Expired OTP Code');
+                }
             }
-
             $otherCustomer = Mage::getModel('customer/customer')->getCollection()
                             ->addAttributeToFilter('mobile_number',array('eq' => $msisdn))
                             ->addAttributeToFilter('entity_id',array('neq' => $data['customer_id']))
@@ -80,11 +81,12 @@ class Bilna_Smsverification_Model_Api2_Verify_Rest_Admin_V1 extends Bilna_Smsver
             return $data;
         } else {
             //Delete old otp failed data of customer and insert new otp failed data
-            $query = "DELETE FROM otp_failed WHERE customer_id=:customer_id AND created_at<:created_at; INSERT INTO otp_failed SET customer_id=:customer_id,otp_code=:otp,created_at=NOW()";
+            $query = "DELETE FROM otp_failed WHERE customer_id=:customer_id AND created_at<:created_at; INSERT INTO otp_failed SET customer_id=:customer_id,otp_code=:otp,created_at=:created_at2";
             $binds = array(
                 'customer_id' => $customerId,
                 'created_at' => $startDate,
-                'otp'   => $data['otp_code']
+                'otp'   => $data['otp_code'],
+                'created_at2' => Mage::getModel('core/date')->date('Y-m-d H:i:s')
             );
             $write->query($query, $binds);
             $this->_critical('Invalid OTP Code');
